Implement message editing functionality and improve conversation handling
- Added support for editing messages, including backend integration for updating message content.
- Introduced new API endpoints for editing messages and retrieving context items.
- Updated conversation data structure to include message IDs for better tracking.
- Enhanced frontend components to manage message editing state and UI interactions.
- Improved caching mechanisms to ensure conversation data is up-to-date after edits.

+++ services/backend/internal/app/agent/prompt.go
@@ -36,7 +36,7 @@ func BuildPlanningPrompt(conn *data.Conn, userID int, query string, contextItems
 						SuggestedQueries: msg.SuggestedQueries,
 						Citations:        msg.Citations,
 						Timestamp:        msg.CreatedAt,
-						TokenCount:       int32(msg.TokenCount),
+						TokenCount:       msg.TokenCount,
 						Status:           msg.Status,
 					}
 					if msg.CompletedAt != nil {
@@ -222,7 +222,7 @@ func BuildFinalResponsePromptWithConversationID(conn *data.Conn, userID int, con
 	return sb.String(), nil
 }
 
-var defaultSystemPromptTokenCount int32
+var defaultSystemPromptTokenCount int
 
 func getDefaultSystemPromptTokenCount(conn *data.Conn) {
 	apiKey, err := conn.GetGeminiKey()
@@ -248,7 +248,7 @@ func getDefaultSystemPromptTokenCount(conn *data.Conn) {
 		return
 	}
 	if CountTokensResponse != nil {
-		defaultSystemPromptTokenCount = CountTokensResponse.TotalTokens
+		defaultSystemPromptTokenCount = int(CountTokensResponse.TotalTokens)
 	}
 }
 
@@ -276,7 +276,7 @@ func BuildPlanningPromptWithConversationID(conn *data.Conn, userID int, conversa
 						SuggestedQueries: msg.SuggestedQueries,
 						Citations:        msg.Citations,
 						Timestamp:        msg.CreatedAt,
-						TokenCount:       int32(msg.TokenCount),
+						TokenCount:       msg.TokenCount,
 						Status:           msg.Status,
 					}
 					if msg.CompletedAt != nil {
